small changes for non-SMAP fitted data

+++ shared/displayerSMAP.m
@@ -184,8 +184,8 @@ end
                         dx=0;
                         dy=dxy;
                      end
-                     px=layers(k).images.srimage.rangex(1)/1000+dx+p.sr_pixrec/1000*5;
-                     py=layers(k).images.srimage.rangey(1)/1000;
+                     px=double(layers(k).images.srimage.rangex(1)/1000+dx+p.sr_pixrec/1000*5);
+                     py=double(layers(k).images.srimage.rangey(1)/1000);
                      lutm=mean(lut,1);lutm=lutm/max(lutm);
                      th=text(p.sr_axes,px,py,p.layernames{k},'Color',lutm,'FontSize',fontsize,'BackgroundColor','k','Units','data');
 %                      th.Units='pixels';
